LayoutTests/accessibility/math-has-non-presentational-children.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<div id="content">
    <div id="math-element" role="math"><mi>sin</mi></div>
</div>

<script>
    description(`
        This test ensures that role="math" elements are not considered to have presentational children.
        This was the case in WAI-ARIA 1.1 (https://www.w3.org/TR/wai-aria-1.1/#math-0), but changed in
        WAI-ARIA 1.2 (https://www.w3.org/TR/wai-aria-1.2/#math).
    `);

    if (window.accessibilityController) {
        var mathElement = accessibilityController.accessibleElementById("math-element");
        shouldBe("mathElement.childrenCount", "1");

        document.getElementById("content").style.visibility = "hidden";
    }
</script>
</body>
</html>
